+++ delphi-frontend/src/main/java/au/com/integradev/delphi/preprocessor/directive/expression/Token.java
@@ -22,7 +22,7 @@ public class Token {
   enum TokenType {
     UNKNOWN,
     INTEGER,
-    DECIMAL,
+    REAL,
     IDENTIFIER,
     STRING,
     EQUALS,
